refactor: move scale defaults to config
